@@ -11,52 +11,51 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.2.14",
-    "@angular/cdk": "~8.2.3",
-    "@angular/common": "~8.2.14",
-    "@angular/compiler": "~8.2.14",
-    "@angular/core": "~8.2.14",
-    "@angular/forms": "~8.2.14",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "~8.2.14",
-    "@angular/platform-browser-dynamic": "~8.2.14",
-    "@angular/router": "~8.2.14",
+    "@angular/animations": "~10.0.5",
+    "@angular/cdk": "~10.1.0",
+    "@angular/common": "~10.0.5",
+    "@angular/compiler": "~10.0.5",
+    "@angular/core": "~10.0.5",
+    "@angular/forms": "~10.0.5",
+    "@angular/material": "^10.1.0",
+    "@angular/platform-browser": "~10.0.5",
+    "@angular/platform-browser-dynamic": "~10.0.5",
+    "@angular/router": "~10.0.5",
     "@mdi/font": "^4.8.95",
-    "@ng-bootstrap/ng-bootstrap": "^5.2.1",
-    "@ngx-translate/core": "^11.0.1",
-    "@zxing/ngx-scanner": "^2.0.1",
-    "angular-archwizard": "^5.0.0",
-    "angularx-qrcode": "^2.1.0",
-    "bootstrap": "^4.4.1",
-    "hammerjs": "^2.0.8",
-    "ngx-infinite-scroll": "^8.0.0",
-    "ngx-material-timepicker": "^5.3.0",
-    "ngx-owl-carousel-o": "^2.0.3",
+    "@ng-bootstrap/ng-bootstrap": "^7.0.0",
+    "@ngx-translate/core": "^13.0.0",
+    "@zxing/ngx-scanner": "^3.0.0",
+    "angular-archwizard": "^6.1.0",
+    "angularx-qrcode": "^10.0.5",
+    "bootstrap": "^4.5.0",
+    "ngx-infinite-scroll": "^9.0.0",
+    "ngx-material-timepicker": "^5.5.3",
+    "ngx-owl-carousel-o": "^3.0.1",
     "ngx-pagination": "^5.0.0",
-    "rxjs": "~6.4.0",
-    "sweetalert2": "^9.7.0",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.9.1"
+    "rxjs": "~6.6.0",
+    "sweetalert2": "^9.17.0",
+    "tslib": "^2.0.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.23",
-    "@angular/cli": "~8.3.23",
-    "@angular/compiler-cli": "~8.2.14",
-    "@angular/language-service": "~8.2.14",
-    "@types/jasmine": "~3.3.8",
+    "@angular-devkit/build-angular": "~0.1000.3",
+    "@angular/cli": "~10.0.3",
+    "@angular/compiler-cli": "~10.0.5",
+    "@angular/language-service": "~10.0.5",
+    "@types/jasmine": "~3.5.11",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "^5.0.0",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.5.3"
+    "@types/node": "^14.0.23",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.1.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~3.3.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~8.10.2",
+    "tslint": "~6.1.0",
+    "typescript": "~3.9.7"
   }
 }
