@@ -2,7 +2,14 @@ import { ModuleWithProviders, NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { MatButtonModule, MatCheckboxModule, MatFormFieldModule, MatInputModule, MatDialogModule, MatCardModule, MatRadioModule, MatProgressSpinnerModule } from '@angular/material';
+import { MatButtonModule } from '@angular/material/button';
+import { MatCardModule } from '@angular/material/card';
+import { MatCheckboxModule } from '@angular/material/checkbox';
+import { MatDialogModule } from '@angular/material/dialog';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { MatRadioModule } from '@angular/material/radio';
 import { TranslateModule } from '@ngx-translate/core';
 
 import { ZXingScannerModule } from '@zxing/ngx-scanner';
@@ -83,7 +90,7 @@ const routes: Routes = [
 })
 
 export class StepperPartnerLoyaltyOfferModule {
-    static forRoot(): ModuleWithProviders {
+    static forRoot(): ModuleWithProviders<StepperPartnerLoyaltyOfferModule> {
         return {
             ngModule: StepperPartnerLoyaltyOfferModule,
             providers: [
