@@ -2,7 +2,13 @@ import { ModuleWithProviders, NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { MatButtonModule, MatCheckboxModule, MatFormFieldModule, MatInputModule, MatDialogModule, MatCardModule, MatTooltipModule } from '@angular/material';
+import { MatButtonModule } from '@angular/material/button';
+import { MatCardModule } from '@angular/material/card';
+import { MatCheckboxModule } from '@angular/material/checkbox';
+import { MatDialogModule } from '@angular/material/dialog';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
+import { MatTooltipModule } from '@angular/material/tooltip';
 import { TranslateModule } from '@ngx-translate/core';
 import { ArchwizardModule } from 'angular-archwizard';
 import { NgxPaginationModule } from 'ngx-pagination';
@@ -118,7 +124,7 @@ import { LanguageSwitcherComponent } from './widgets/language-switcher/language-
 })
 
 export class ViewsModule {
-    static forRoot(): ModuleWithProviders {
+    static forRoot(): ModuleWithProviders<ViewsModule> {
         return {
             ngModule: ViewsModule,
             providers: [
