@@ -26,8 +26,8 @@ import { GeneralList } from 'src/app/core/interfaces/general-list.interface';
 })
 
 export class EditEventComponent implements OnInit {
-  @ViewChild('remove_item', { static: false }) remove_item: NgbModalRef;
-  @ViewChild('fileInput', { static: false }) imageInput: ElementRef;
+  @ViewChild('remove_item') remove_item: NgbModalRef;
+  @ViewChild('fileInput') imageInput: ElementRef;
 
 	/**
 	 * Configuration and Static Data
